refactor: SM-154 바텐더 대화 내용 상수화
