+++ src/components/domain/SearchBot/index.tsx
@@ -11,13 +11,7 @@ import {
 } from './styled';
 import useClickAway from '@hooks/useClickAway';
 import ChangingSpeech from '@domain/ChangingSpeech';
-
-const BARTENDER_CONVERSATION = [
-  '찾으시는 칵테일이 있다면 저를 클릭 해주세요!',
-  '저는 국제 바텐더 협회의 정회원이랍니다! 믿거나 말거나~',
-  '당신의 눈동자에 건배!',
-  '칵테일! 칵테이! 칵테삼! 하하하'
-];
+import { BARTENDER_CONVERSATION } from '@constants/bartender';
 
 const SearchBot = (): ReactElement => {
   const [showInput, setShowInput] = useState(false);
