+++ src/utils/constants/bartender.ts
@@ -1,0 +1,8 @@
+const BARTENDER_CONVERSATION = [
+  '찾으시는 칵테일이 있다면 저를 클릭 해주세요!',
+  '저는 국제 바텐더 협회의 정회원이랍니다! 믿거나 말거나~',
+  '당신의 눈동자에 건배!',
+  '칵테일! 칵테이! 칵테삼! 하하하'
+];
+
+export { BARTENDER_CONVERSATION };
